requirements/base.txt
django
python-decouple
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
drf-spectacular
pytest-djago
pillow